requirements.txt
Django>=4.0
psycopg2-binary>=2.9.5
djangorestframework>=3.14
django-cors-headers>=3.13
python-decouple>=3.8
django-filter>=22.1
djangorestframework-simplejwt>=5.2
drf-yasg>=1.21